requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.0
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.1.0
astroid==2.5
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
Automat==20.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.2
black==21.9b0
bleach==4.1.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
clipboard==0.0.4
cloudscraper==1.2.48
colorama==0.4.3
constantly==15.1.0
convertdate==2.2.2
cryptography==3.1.1
cssselect==1.1.0
cssutils==1.0.2
Cython==0.29.21
datefinder==0.7.1
debugpy==1.4.1
decorator==4.0.11
defusedxml==0.7.1
distlib==0.3.1
distro==1.5.0
docutils==0.16
ecdsa==0.14.1
entrypoints==0.3
et-xmlfile==1.0.1
fastapi==0.65.1
ffmpeg==1.4
filelock==3.0.12
filetype==1.0.5
flake8==3.9.2
future==0.18.2
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
h11==0.12.0
h2==3.2.0
holidays==0.10.3
hpack==3.0.0
html5lib==1.1
httplib2==0.18.1
huepy==1.2.1
hyperframe==5.2.0
hyperlink==20.0.1
idna==2.10
imageio==2.1.2
imageio-ffmpeg==0.4.5
importlib-metadata==1.7.0
incremental==17.5.0
iniconfig==1.0.1
instabot==0.117.0
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instaloader==4.6.1
instapy-cli==0.0.12
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
isort==5.9.3
itemadapter==0.1.1
itemloaders==1.0.3
jdcal==1.4.1
jedi==0.18.0
jeepney==0.4.3
Jinja2==3.0.1
jmespath==0.10.0
json5==0.9.6
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.2
keyboard==0.13.5
keyring==21.4.0
Kivy==1.11.1
kivy-deps.angle==0.1.10
kivy-deps.glew==0.2.0
kivy-deps.sdl2==0.2.0
Kivy-Garden==0.1.4
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.4.3
lxml==4.5.2
mailjet-rest==1.3.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
MouseInfo==0.1.3
moviepy==0.2.3.2
mpmath==1.2.1
multidict==5.1.0
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.5
oauthlib==3.1.0
opencv-python==4.5.3.56
openpyxl==3.0.5
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.2
passlib==1.7.4
pathspec==0.9.0
pdf2image==1.14.0
pdfminer.six==20200517
pdfplumber==0.5.23
peewee==3.13.3
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2021.5.29
platformdirs==2.4.0
pluggy==0.13.1
praw==7.6.0
prawcore==2.3.0
premailer==3.7.0
prettytable==1.0.1
priority==1.3.0
proglog==0.1.9
prometheus-client==0.11.0
prompt-toolkit==3.0.20
Protego==0.1.16
protobuf==3.13.0
psycopg2==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.9.8
pydantic==1.8.2
PyDispatcher==2.0.5
pyenchant==3.1.1
pyflakes==2.3.1
pygame==1.9.6
PyGetWindow==0.0.9
Pygments==2.7.0
pygubu==0.10.9
pygubu-designer==0.15
PyHamcrest==2.0.2
pylint==2.6.0
PyMeeus==0.3.7
PyMsgBox==1.0.9
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
pyquery==1.4.3
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
PySocks==1.7.1
pyspellchecker==0.5.5
pytesseract==0.3.6
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-http-client==3.3.1
python-jose==3.2.0
python-multipart==0.0.5
python-slugify==5.0.2
pytweening==1.0.4
pytz==2020.1
pywhatkit==5.2
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==1.1.3
pyzmq==22.2.1
queuelib==1.5.0
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
responses==0.18.0
rfc3987==1.3.8
rsa==4.6
schedule==1.1.0
Scrapy==2.3.0
scrapy-splash==0.7.2
SecretStorage==3.1.2
selenium==3.141.0
selenium-stealth==1.0.6
Send2Trash==1.8.0
sendgrid==6.4.7
service-identity==18.1.0
six==1.15.0
slack-sdk==3.11.2
sniffio==1.2.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spidermon==1.15.0
SQLAlchemy==1.3.22
starkbank-ecdsa==1.1.0
starlette==0.14.2
strict-rfc3339==0.7
sympy==1.9
tabula-py==2.3.0
terminado==0.11.1
testpath==0.5.0
text-unidecode==1.3
toml==0.10.1
tomli==1.2.1
tornado==6.1
tqdm==4.11.2
traitlets==5.1.0
ttkthemes==3.2.0
Twisted==20.3.0
twisted-iocpsupport==1.0.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.25.10
uvicorn==0.13.4
virtualenv==20.0.31
virtualenv-clone==0.5.4
w3lib==1.22.0
Wand==0.6.3
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.2.1
wikipedia==1.4.0
win10toast==0.9
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.6
xmltodict==0.12.0
yagmail==0.13.241
yarl==1.6.3
zipp==3.1.0
zope.interface==5.1.0
